+++ apps/backoffice-v2/src/pages/Entity/components/Case/hooks/usePendingRevisionEvents/usePendingRevisionEvents.tsx
@@ -1,61 +1,31 @@
 import { TWorkflowById } from '@/domains/workflows/fetchers';
-import { useCallback, useMemo } from 'react';
-import { calculateAllWorkflowPendingEvents } from '@/pages/Entity/components/Case/hooks/usePendingRevisionEvents/utils/calculate-pending-workflow-events';
-import { CommonWorkflowEvent, CommonWorkflowStates } from '@ballerine/common';
+import { useCallback } from 'react';
+import { CommonWorkflowEvent } from '@ballerine/common';
 import { checkIsKybExampleVariant } from '@/lib/blocks/variants/variant-checkers';
 import { useRevisionCaseMutation } from '@/domains/workflows/hooks/mutations/useRevisionCaseMutation/useRevisionCaseMutation';
-import { IPendingEvent } from './interfaces';
-
-const composeUniqueWorkflowEvents = (
-  acc: Record<string, IPendingEvent>,
-  pendingWorkflowEvent: IPendingEvent,
-) => {
-  acc[`${pendingWorkflowEvent?.workflowId}-${pendingWorkflowEvent?.eventName}`] =
-    pendingWorkflowEvent;
-
-  return acc;
-};
-
-const isPendingEventIsRevision = (pendingWorkflowEvent: IPendingEvent) =>
-  pendingWorkflowEvent?.eventName === CommonWorkflowEvent.REVISION ||
-  pendingWorkflowEvent?.workflowState === CommonWorkflowStates.MANUAL_REVIEW;
 
 export const usePendingRevisionEvents = (
   mutateRevisionCase: ReturnType<typeof useRevisionCaseMutation>['mutate'],
   workflow?: TWorkflowById,
 ) => {
-  const pendingWorkflowEvents = useMemo(() => {
-    if (!workflow) return;
-
-    return calculateAllWorkflowPendingEvents(workflow);
-  }, [workflow]);
-
   const onMutateRevisionCase = useCallback(() => {
-    if (!pendingWorkflowEvents || !workflow) return;
+    if (!workflow?.nextEvents?.some(nextEvent => nextEvent === CommonWorkflowEvent.REVISION)) {
+      return;
+    }
 
-    const uniqueWorkflowEvents = pendingWorkflowEvents
-      .filter((pendingWorkflowEvent: IPendingEvent) =>
-        isPendingEventIsRevision(pendingWorkflowEvent),
-      )
-      .reduce(
-        (acc: Record<string, IPendingEvent>, pendingWorkflowEvent: IPendingEvent) =>
-          composeUniqueWorkflowEvents(acc, pendingWorkflowEvent),
-        {},
-      );
+    mutateRevisionCase({ workflowId: workflow?.id });
 
-    Object.keys(uniqueWorkflowEvents).forEach(pendingWorkflowKeys => {
-      const pendingWorkflowEvent = uniqueWorkflowEvents[pendingWorkflowKeys];
-      mutateRevisionCase({ workflowId: pendingWorkflowEvent!.workflowId });
+    const isKybExampleVariant = checkIsKybExampleVariant(workflow?.workflowDefinition);
 
-      const isKybExampleVariant = checkIsKybExampleVariant(workflow.workflowDefinition);
-      if (!isKybExampleVariant) return;
+    if (!isKybExampleVariant) {
+      return;
+    }
 
-      window.open(
-        `${workflow?.context?.metadata?.collectionFlowUrl}/?token=${pendingWorkflowEvent?.token}`,
-        pendingWorkflowEvent?.token,
-      );
-    });
-  }, [mutateRevisionCase, pendingWorkflowEvents, workflow]);
+    window.open(
+      `${workflow?.context?.metadata?.collectionFlowUrl}/?token=${workflow?.context?.metadata?.token}`,
+      workflow?.context?.metadata?.token,
+    );
+  }, [mutateRevisionCase, workflow]);
 
-  return { onMutateRevisionCase, pendingWorkflowEvents };
+  return { onMutateRevisionCase };
 };