+++ examples/headless-example/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @ballerine/headless-example
 
+## 0.3.102
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.9.82
+  - @ballerine/workflow-browser-sdk@0.6.103
+
+## 0.3.101
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.81
+  - @ballerine/workflow-browser-sdk@0.6.102
+
+## 0.3.100
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.9.80
+  - @ballerine/workflow-browser-sdk@0.6.101
+
 ## 0.3.99
 
 ### Patch Changes