+++ examples/report-generation-example/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @ballerine/report-generation-example
 
+## 0.2.35
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/react-pdf-toolkit@1.2.91
+
+## 0.2.34
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/react-pdf-toolkit@1.2.90
+
+## 0.2.33
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/react-pdf-toolkit@1.2.89
+
 ## 0.2.32
 
 ### Patch Changes