+++ packages/react-pdf-toolkit/CHANGELOG.md
@@ -1,5 +1,60 @@
 # @ballerine/react-pdf-toolkit
 
+## 1.2.91
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/config@1.1.36
+  - @ballerine/ui@0.7.117
+
+## 1.2.90
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/config@1.1.35
+  - @ballerine/ui@0.7.116
+
+## 1.2.89
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/config@1.1.34
+  - @ballerine/ui@0.7.115
+
+## 1.2.88
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/ui@0.7.114
+
+## 1.2.87
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/ui@0.7.113
+
+## 1.2.86
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/ui@0.7.112
+
+## 1.2.85
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/ui@0.7.111
+
 ## 1.2.84
 
 ### Patch Changes