+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,53 @@
 # @ballerine/ui
 
+## 0.7.117
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.9.82
+
+## 0.7.116
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.81
+
+## 0.7.115
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.9.80
+
+## 0.7.114
+
+### Patch Changes
+
+- Bump
+
+## 0.7.113
+
+### Patch Changes
+
+- Bump
+
+## 0.7.112
+
+### Patch Changes
+
+- Bump UI & KYB
+
+## 0.7.111
+
+### Patch Changes
+
+- Added minimumAge validator
+
 ## 0.7.110
 
 ### Patch Changes